backend/requirements.txt
fastapi>=0.129.0
uvicorn[standard]>=0.41.0
python-multipart>=0.0.22
python-dotenv>=1.2.1
google-genai>=1.64.0
google-cloud-aiplatform>=1.138.0
supabase>=2.28.0
stripe>=14.3.0
pydantic>=2.11.0
pydantic-settings>=2.13.1
httpx>=0.28.1